Colenso/requirements.txt
Django==1.9.4
argparse==1.2.1
django-bootstrap3==7.0.1
hashids==1.1.0
lxml==3.5.0
pycrypto==2.6.1
requests==2.9.1
simple-crypt==4.1.7
wsgiref==0.1.2
